+++ imessage-database/src/util/typedstream/tests/mod.rs
@@ -1,0 +1,3 @@
+mod test_bubble;
+mod test_parser;
+mod test_type;
